+++ nimjl/sugar/dicttuples.nim
@@ -1,9 +1,10 @@
-import ./coretypes
-import ./modfuncs
+import ../cores
+import ../types
+import ../functions
 
 import std/tables
 import std/strutils
-
+import std/strformat
 
 proc jlTupleToNim*(val: JlValue, tup: var tuple) =
   # collect(keys(val))
@@ -44,13 +45,34 @@ proc jlDictToNim*[U, V: string|SomeNumber|bool](val: JlValue, tab: var Table[U, 
     tab[key.to(U)] = val.to(V)
 
 # Recursive import strategy
-import converttypes
+import ./converttypes
 # Tuple helpers -> result is memory managed by Julia's GC
 # Convert object as tuple ?
-proc nimToJlTuple*(v: tuple|object): JlValue =
+proc nimToUnnamedJlTuple(v: tuple): JlValue =
+  var tupStr = $v
+  result = jlEval(tupStr)
+
+proc nimToNamedJlTuple(v: tuple|object): JlValue =
   result = jlEval("NamedTuple()")
   for name, field in v.fieldPairs:
     result = jlCall(JlBase, "setindex", result, toJlVal(field), jlSym(name))
+
+proc isNamedTuple(v: tuple) : bool =
+  var i = 0
+  for name, field in v.fieldPairs:
+    if name != &"Field{i}":
+      return true
+    inc(i)
+  return false
+
+proc nimToJlTuple*(v: tuple): JlValue =
+  if isNamedTuple(v):
+    nimToNamedJlTuple(v)
+  else:
+    nimToUnnamedJlTuple(v)
+
+proc nimToJlTuple*(v: object): JlValue =
+  nimToNamedJlTuple(v)
 
 proc nimTableToJlDict*[U, V: string|SomeNumber](tab: Table[U, V]): JlValue =
   result = jlEval("Dict()")