+++ tests/test2.jl
@@ -1,3 +1,0 @@
-function example()
-  println(hello)
-end